requirements.txt
cryptography
djangorestframework
djangorestframework-simplejwt
markdown
